convert steward to using "Ihrzen"
The steward is a special case in the german translation. She is the only
character to use "Sie" ("Siezen"), which feels strange. Convert her to
use "Ihrzen", which is a little bit more common in the context of a
middle ages steward talking to his liege (in a fantasy world). The
english original has no indication whatsoever, that the steward talks in
any way different than any other character.